refactor(map): move webhook handlers
